Filter feed requests by date range (relative and absolute), refactoring feed search logic by moving it to the model, refactoring search by date logic as a helper used by both the Feed model and the CheckSearch library
