+++ app/models/feed.rb
@@ -1,4 +1,6 @@
 class Feed < ApplicationRecord
+  include SearchHelper
+
   check_settings
 
   has_many :requests
@@ -54,6 +56,35 @@ class Feed < ApplicationRecord
   def item_belongs_to_feed?(pm)
     items = self.project_media_ids(pm.team_id)
     items.include?(pm.id)
+  end
+
+  def search(args = {})
+    request_id = (args['request_id'].to_i == 0 ? nil : args['request_id'].to_i)
+
+    query = Request.where(request_id: request_id, feed_id: self.id)
+    query = query.or(Request.where(id: request_id, feed_id: self.id)) unless request_id.nil?
+
+    # Filters
+    query = query.where('medias_count >= ?', args['medias_count_min'].to_i) unless args['medias_count_min'].blank?
+    query = query.where('medias_count <= ?', args['medias_count_max'].to_i) unless args['medias_count_max'].blank?
+    query = query.where('requests_count >= ?', args['requests_count_min'].to_i) unless args['requests_count_min'].blank?
+    query = query.where('requests_count <= ?', args['requests_count_max'].to_i) unless args['requests_count_max'].blank?
+    query = query.where('requests.created_at' => Range.new(*format_times_search_range_filter(JSON.parse(args['request_created_at']), nil))) unless args['request_created_at'].blank?
+
+    # Sort
+    sort = {
+      'requests' => 'requests_count',
+      'medias' => 'medias_count',
+      'last_submitted' => 'last_submitted_at',
+      'subscriptions' => 'subscriptions_count',
+      'media_type' => 'medias.type',
+      'fact_checked_by' => 'fact_checked_by_count',
+      'fact_checks' => 'project_medias_count'
+    }[args['sort'].to_s] || 'last_submitted_at'
+    sort_type = args['sort_type'].to_s.downcase == 'asc' ? 'ASC' : 'DESC'
+    query = query.joins(:media) if sort == 'medias.type'
+
+    query.order(sort => sort_type).offset(args['offset'].to_i)
   end
 
   # This takes some time to run because it involves external HTTP requests and writes to the database:
